Statistically compare interactions between groups of samples

Statistically compare interactions between groups of samples

compare_group_interactions(
  seurat_object,
  interaction_stats,
  sample_replicates,
  sample_groups,
  metadata_grouping
)

Arguments

seurat_object
Seurat object containing expression data acorss all samples and cells

interaction_stats
Tibble from celltalk function, usually filtered to the top significant ligand and receptor interactions of interest

sample_replicates
Name of the meta.data column in a Seurat object that has the samples of the individual replicate samples

sample_groups
Name of the meta.data column in a Seurat object that has the name of the sample group

metadata_grouping
Name of the meta.data column in a Seurat object that has the name of the groups of cells to evaluate (e.g. "cell_types" containing previously identified cell types)

Value

A list of linear models containing the coefficients and statistics for assessing differences in joint ligand and receptor expression between sample groups. Note that this requires replicate samples from each samples group.