Assess joint means between scrambled interaction pairs for ligands and receptors across cell types

Assess joint means between scrambled interaction pairs for ligands and receptors across cell types

ligand_receptor_scramble(
  ligand_receptor_means,
  ligand_mean_dataframe,
  receptor_mean_dataframe,
  ligand_receptor_pairs,
  number_cell_types,
  scramble_times
)

Arguments

ligand_receptor_means
Tibble generated by ligand_receptor_means function. Used to generate create a joint expression weighting for ligand receptor pairs.

ligand_mean_dataframe
Mean expression values for ligands across cells from a given condition

receptor_mean_dataframe
Mean expression values for receptors across cells from a given condition

ligand_receptor_pairs
Data.frame of ligands, receptors and interactions in the format of ramilowski_pairs provided by this package

number_cell_types
The number of unique cell types in a given sample or condition. This will be used to assess ligand and receptor interactions across cell types.

scramble_times
Number of times to random pair ligands and receptors to generate a background distribution.

Value

Generates a tibble containing joint resulting from scrambling ligand and receptor pairs.